# openapi.json
{
  "components": {
    "schemas": {
...
      "CalendarStatusOut": {
        "properties": {
          "account_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Email"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "connected": {
            "title": "Connected",
            "type": "boolean"
          },
          "failed": {
            "default": 0,
            "title": "Failed",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "pending": {
            "default": 0,
            "title": "Pending",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "required": [
          "provider",
          "label",
          "available",
          "connected"
        ],
        "title": "CalendarStatusOut",
        "type": "object"
      },
...
      "ConnectOut": {
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "ConnectOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SyncOut": {
        "properties": {
          "changed": {
            "title": "Changed",
            "type": "integer"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "pushed": {
            "title": "Pushed",
            "type": "integer"
          }
        },
        "required": [
          "pushed",
          "changed"
        ],
        "title": "SyncOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Production-ready FastAPI backend supporting user authorization, lead records, and calls metrics with a PostgreSQL database.",
    "title": "aVn Agent API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/inbound/end": {
      "post": {
        "description": "End a live inbound call by closing its LiveKit room.",
        "operationId": "end_inbound_call_api_inbound_end_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "End Inbound Call",
        "tags": [
          "Inbound Management"
        ]
      }
    },
...
    "/api/integrations/calendar": {
      "get": {
        "description": "Which calendars this workspace has connected, and how their sync is going.",
        "operationId": "calendar_status_api_integrations_calendar_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CalendarStatusOut"
                  },
                  "title": "Response Calendar Status Api Integrations Calendar Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Calendar Status",
        "tags": [
          "Calendar sync"
        ]
      }
    },
    "/api/integrations/calendar/{provider}": {
      "delete": {
        "description": "Stop syncing and revoke aVn's access. Events already in the calendar stay.",
        "operationId": "disconnect_calendar_api_integrations_calendar__provider__delete",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Disconnect Calendar",
        "tags": [
          "Calendar sync"
        ]
      }
    },
    "/api/integrations/calendar/{provider}/connect": {
      "post": {
        "description": "The provider's sign-in page. After consent it returns to /callback, which stores the connection.",
        "operationId": "start_connect_api_integrations_calendar__provider__connect_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Start Connect",
        "tags": [
          "Calendar sync"
        ]
      }
    },
    "/api/integrations/calendar/{provider}/sync": {
      "post": {
        "description": "Retry failed events and read changes now instead of waiting for the next automatic run.",
        "operationId": "sync_calendar_api_integrations_calendar__provider__sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Sync Calendar",
        "tags": [
          "Calendar sync"
        ]
      }
    },
...
    "/api/outbound/end": {
      "post": {
        "description": "End a live outbound call by closing its LiveKit room.",
        "operationId": "end_outbound_call_api_outbound_end_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "End Outbound Call",
        "tags": [
          "Call Management"
        ]
      }
    },
...
  }
}